Only run the deployment failure step if needed.
It was still executing, for example, when a step prior to the create deployment step had failed.
In which case it was failing with:

Error: .github/workflows/docs.yml (Line: 143, Col: 23):
Error: The template is not valid. .github/workflows/docs.yml (Line: 143, Col: 23): Error reading JToken from JsonReader. Path '', line 0, position 0.

So, now, "needed" is satisfied by testing all of the following are satisfied:
- did we create a deployment? YES
- did we succeed that deployment? NO
- has a step failed? YES