If we fail to enter presence due to token capabilities, fetch new token and retry
This is taken from the corresponding feature on Android
(https://github.com/ably/ably-asset-tracking-android/pull/703).

The use case in mind is that of an application where users only have
permission to interact with the channels that correspond to a certain
set of trackables, where this permission is enforced by a service
backend that vends Ably tokens whose capabilities only allow access to
the channels that correspond to the current allowed set of trackables.

If at some point a new trackable is added to a user’s allowed set, the
initial attempt to enter presence will fail since we’re using an Ably
token whose capabilities don’t include the current trackable. So we need
to fetch a new Ably token (which hopefully will now include permissions
for the current trackable) and try to enter presence again.

Closes #311.