Wrap Ably SDK in a protocol and inject into DefaultAbly
I want to be able to write unit tests for the behaviour of DefaultAbly,
without having to interact with the ably-cocoa SDK.

I originally wrote these protocols hoping that I’d be able to make
ably-cocoa’s classes directly conform to them, e.g.

extension ARTRealtime: AblySDKRealtime {}

This _mainly_ worked, but there were a few protocol methods that,
although implemented with a seemingly identical signature in the
ably-cocoa classes, the compiler refused to recognise as being
implemented. So I eventually gave up on this approach and have wrapped
everything inside the AblyCocoaSDK* types.