Move LogHandler+Default extension from Core to Internal
This is only needed internally; it doesn’t need to be part of the
external API of the Asset Tracking SDKs. Users of the SDKs shouldn’t be
interacting with the methods of the LogHandler interface; if they wish
to emit logs then they should do through their own API.

I know that in Android they are technically part of the public API
(LoggingExtensions.kt in c52a9a5), but I’m not sure it’s intentional,
and I don’t think it’s a decision we need to align with anyway.