Do not repeat work when we try to add a trackable twice
This handles the case where a user adds a trackable and then tries to
add another trackable with the same ID before the first add operation
has completed. It ensures that we do not repeat the work of joining Ably
presence, instead saving the completion handler for the second add
operation and calling it when the first add operation completes.

This is copied from the Android implementation (as of 3a0d833).

It’s worth pointing out that this continues our existing notion of
trackables being “the same” when adding them – that is, that we only
compare them by ID. This might not be ideal and we’ve already opened
issue #357 to revisit this.

Closes #341.