Fix some incorrect conversions between milliseconds and seconds in the subscriber example app
We had assumed that things that were seconds were milliseconds and
vice-versa.

I think things would be easier if we just used TimeInterval in our
public API to represent durations, but I’ve created #378 for that.

The updated usage example in the README is taken from the
SubscriberExample app.

I discovered this whilst investigating #375 (when using animation, the
asset position on map quickly drifts far from received asset positions),
but this change alone doesn’t fix that issue.