Calculate animation frames on demand
The existing approach was to calculate the map marker position for each
animation frame upfront, assuming a constant frame rate.

This had several shortcomings:

- #317: It’s incorrect to assume a constant frame rate, since the frame
  rate can vary for reasons such as the main thread being blocked or the
  system choosing to dynamically change the frame rate.

- #375: If new locations to be animated arrive when there are still
  locations waiting to be animated, we should shorten the duration of
  the animation of the current locations so that we maintain a constant
  animation duration.

The new approach in this commit changes things so that we calculate the
map marker position one frame at a time, taking into account the current
frame rate and number of locations waiting to be animated.

It moves the bulk of the calculations into a separate class for the sake
of testability, and adds unit tests for this class.

I am also satisfied, after this commit and the previous e0badb8 and
637b814, that it is not possible for the animation progress to exceed
100%, hence closing #369.

Closes #317, closes #369, closes #375.