Expose information about publisher presence to subscriber
This copies the feature added in ably/ably-asset-tracking-android#754.

I don’t fully understand the motivation for this feature - what I found
was https://github.com/ably/ably-asset-tracking-common/issues/58:

> Right now the TrackableState is calculated from 3 parts: Ably
connection state, trackable channel connection state, and the presence
of the publisher on the channel. While it should be enough in most cases
to just know the final trackable state, in others it might be good to
know the state and each individual part of it. For example, the
subscriber users might stop it when the publisher is not present for a
long time since the start.

Closes #368.