Add logging of locations received from Mapbox
I’m doing this as part of investigating #377 (repeated locations being
published). It seems to be like it’s an issue with the locations being
received from Mapbox, so I want some further logging to aid my
investigation.

This first step of logging is very simple, and isn’t actually that
helpful, since the default String(reflecting:_) description of
CLLocation is like this:

"<+37.63853126,-122.44089953> +/- 5.00m (speed 33.29 mps / course 325.90) @ 10/4/22, 12:30:15 PM Brasilia Standard Time"

Which doesn’t show all of the values in the location, nor its precise
timestamp, so it’s hard to compare them over time. Will improve on this.