Set NavigationSettings’s statusUpdatingSettings to some magic value 🤷
This is taken from commit b14e3b0 in #381. Would love to have some more
justification. Seems to fix #377.

Seems to have been added in Android commit b26a2ed:

> InternalUtils.setUnconditionalPollingPatience(Long.MAX_VALUE)

And then a bit more of an explanation added in 8ddd20d. But the
explanation is just the following:

> Explanation from the Mapbox team: "By disabling this setting we strive
> for predictability and stability"

The documentation for Android’s setUnconditionalPollingPatience method
says the following:

> /**
>  * Internal API used for testing. Sets the static unconditional polling patience value for all
>  * instances in the process. Needs to be set prior to [MapboxNavigation] creation.
>  *
>  * Pass `null` to reset to default.
>  *
>  * Do not use in a production environment.
>  */
> fun setUnconditionalPollingPatience(patienceInMillis: Long?) {
>     UNCONDITIONAL_POLLING_PATIENCE_MILLISECONDS = patienceInMillis ?: 2000L
> }