Set NavigationSettings’s statusUpdatingSettings to some magic value 🤷
This seems to fix #377, where we were seeing the Mapbox SDK
emit the same location multiple times with the same timestamp, as
described by this comment:
https://github.com/ably/ably-asset-tracking-swift/issues/377#issuecomment-1268576416

To summarise that comment, the problem was reproducible by running the
PublisherExampleSwiftUI example app with .trace level logging enabled,
and setting the "Freeway Drive" simulated location in the iOS Simulator.
Grepping through the generated logs for "Differences between location
and last received location", we saw that Mapbox would emit an enhanced
location update followed a second or two later by another enhanced
location update which was identical except in timestamp. It would then
eventually followed by an enhanced location update with updated
coordinate / course / speed, and the pattern would then repeat. This
gave the impression that the publisher device was moving, then stopping
for a while, then moving again, then stopping and so on. This was then
reflected in the map animation displayed by the subscriber, which didn’t
look very good.

With the change in this commit, that no longer occurs - every enhanced
location update emitted by Mapbox now has a different coordinate to the
previous one. And as a result, the subscriber map animation is now
continuously-moving.

This is taken from commit b14e3b0 in #381. I’m waiting to find out where
this code came from. Would love to have some more justification. And
then I’m also not sure why it fixes this issue.

Seems to have been added in Android commit b26a2ed:

> InternalUtils.setUnconditionalPollingPatience(Long.MAX_VALUE)

And then a bit more of an explanation added in 8ddd20d. But the
explanation is just the following:

> Explanation from the Mapbox team: "By disabling this setting we strive
> for predictability and stability"

The documentation for Android’s setUnconditionalPollingPatience method
says the following:

> /**
>  * Internal API used for testing. Sets the static unconditional polling patience value for all
>  * instances in the process. Needs to be set prior to [MapboxNavigation] creation.
>  *
>  * Pass `null` to reset to default.
>  *
>  * Do not use in a production environment.
>  */
> fun setUnconditionalPollingPatience(patienceInMillis: Long?) {
>     UNCONDITIONAL_POLLING_PATIENCE_MILLISECONDS = patienceInMillis ?: 2000L
> }