Publisher example: Use a dummy LocationManager for view previews
This gives us control over the data shown in the previews. The immediate
improvement here is that the map region displayed in the previews now
shows a city view instead of the middle of nowhere.

I don’t know if there’s an easy way to not have to include the
PreviewLocationManager code in the app bundle, but will just live with
it for now.