Publisher example: Add a publisher details screen
We introduce a new view, PublisherDetailsView, which has a button for
adding a trackable (we use what was MainView, renamed to
AddTrackableView).

PublisherDetailsView also displays a list of the publisher’s trackables,
and links from each trackable to the map view. The map view does not
yet make use of the publisher that we’ve already created, and still
creates its own publisher. This is the next thing we’ll address.