Publisher example: Move app-wide settings resolution to a Settings tab
This keeps the settings that relate to the creation of the publisher
(whether to publish raw locations, constant resolution, vehicle and
routing profiles) in PublisherSettingsView. But it moves the other
settings (just the default resolution the app should use when creating a
trackable, and whether to display a Mapbox map) to a new view
SettingsView, which we display in a “Settings” tab.

I haven't updated this to dynamically switch which map is displayed if
the user changes the toggle; the change will only take effect the next
time they create a map view.