Publisher example: Add button on publisher details view to change routing profile
This doesn't seem to be working though – the invocation of
routeProvider.changeRoutingProfile on line 278 of DefaultPublisher.swift
in the SDK never seems to call its callback. This seems to be happening
on `main` too; I don’t think it’s something I’ve broken in this branch.
I’ve created issue #420 for this.

The UI code for changing the routing profile is taken from MapView, from
where I’ll shortly be removing it.