Remove Sources/AblyAssetTrackingCore/Core.xcodeproj
I don’t know what this is, don’t understand why it would be there, and
by symmetry with the other packages in Sources, don’t think it should be
there. The commit history doesn’t really explain its presence, so I’m
going to remove it and see if anything bad happens.