Publisher example: Add a client for fetching list of LocationHistoryData files from S3
We use Amazon’s Amplify client library for this, same as the Android
example app, so that we can use the same configuration file for both
platforms.

The functionality of this class will only be available if the user
places an Amplify configuration file at
Examples/PublisherExampleSwiftUI/PublisherExampleSwiftUI/Optional Resources/amplifyconfiguration.json.
I’ll document this when I add the UI side of things.

Closes #411, closes #436.

Co-authored-by: Lawrence Forooghian <lawrence@forooghian.com>