Add LocationHistoryData type and ability to use it as a location source
This is based on the Android API (as of commit 8016b0f there), without
the onDataEnded callback, which we don’t have a need for at the moment
and can add if needed in the future.

The LocationHistoryData type also matches the
location-history-data-schema.json JSON schema in commit 5337ae of
ably/ably-asset-tracking-common.

Closes #290.

Co-authored-by: Artur Jaworski <a.jaworski@nomtek.com>