Record location history data and expose it in publisher API
The implementation is roughly taken from Mapbox.kt of the Android asset
tracking codebase at d61ca7f.

This increases the complexity of DefaultLocationService, and I would
have liked to have written tests for this new functionality, but the
class currently has no tests and it would require a chunk of work to
make it testable (we need to mock out Mapbox). So I’ve created a
separate issue #456 for writing tests for this class.

Closes #111.