Publisher example: Make sure we start publisher from main thread
I noticed that in the publisher example app, the trackables are no
longer transitioning to the online state. This appears to have been
broken by a change in ac20806 where we started creating the publisher
inside an async function and hence potentially no longer on the main
thread.

This SDK does not explain the rules re which from which threads it can
be interacted with. I have created #445 to decide and document these
rules.