Fix bug where, after removing a trackable, can’t add another with same ID
This occurs because the second time around, the channel will be in a
DETACHED state, and trying to enter presence on that channel will give
the error "unable to enter presence channel (incompatible channel state:
Detached)".

So, we follow Android’s approach [1], and check the channel’s state and
attach if necessary before we try to enter presence.

Closes #431.

[1] https://github.com/ably/ably-asset-tracking-android/blob/d61ca7f3a1dc2f8c3ed5eb5def21ee9a3e0229b8/common/src/main/java/com/ably/tracking/common/Ably.kt#L363-L365