Handle versioning of archived LocationHistoryData data
Currently emitting a "wip" prerelease modifier since we don’t yet emit
the `agents` property (we’ll do this in #451).

This is done with reference to the version-1 and version-2.0.0 schemas
in the external/common repo.

We deprecate the `version` instance property on LocationHistoryData,
since it’s not really a property of the object, but rather of the way in
which the current version of the library will serialize that object.
I’ve replaced it with a class property called `archiveVersion` which
will return the value that will be placed in the archived data’s
`version` field (so that we can continue to use it in the filename of
the files that we upload in the publisher example app).

In addition to (or perhaps replacing) the tests for Encodable
conformance that I’ve added, I’d like to validate the JSON-encoded
LocationHistoryData objects against the schemas in the external/common
repo. I’ve created #470 for doing this (throughout the library, not just
for this type).

Closes #466.