Turn off library evolution mode in assemble.sh
Library evolution mode [1] is designed for people who distribute binary
frameworks (which we don’t). It imposes restrictions on the code you can
write.

I’m not sure why we turned it on in the first place; it’s not explained
in 7d52c8f nor its containing PR #213. I see no reason for us to use it.

This came to my attention because I was trying to add a dependency on
the Version package [2] and was getting strange errors. I noticed that
turning off library evolution mode stopped these errors. Since we have
no need for library evolution mode, let’s just turn it off.

[1] https://www.swift.org/blog/library-evolution/
[2] https://github.com/mxcl/Version