Add TemporaryFile class
I’ll be adding a commit next where the SDK emits the location of a file
that the user may wish to access. I want to make sure that this file
gets cleaned up without the user having to do anything, but also that
the file only gets deleted once the user has finished with it.