Keep track of which publisher callbacks need to be called on main thread
The aim here is to make sure that we call a callback on the main thread
if and only if it originates from the public API of DefaultPublisher.

The way that this is implemented allows us to, inside DefaultPublisher,
treat all callbacks the same, whether they originate from the public API
or internally in the class, whilst still being sure that the threading
contract of the public API will be fulfilled.

The motivation here was that we have a performAddOrTrack method which
sometimes calls the callbacks on the main thread and sometimes calls
them directly, but it’s not very clear how that method got the knowledge
about when it needs to perform the callback on the main thread.