Log info about which subsystem emitted message
This adds a tag like [assetTracking.subsystem1.subsystem2] to each log
message, identifying which subsystem (which might be a type, or might be
the ably-cocoa SDK) emitted the message.

For example, a log from ably-cocoa inside the Publisher SDK would be
tagged [assetTracking.publisher.DefaultAbly.ablySDK]. (This might turn
out to be a little too verbose, but we can tweak that later.)

This will allow us to always identify the output of the ably-cocoa SDK,
and will also allow us to remove all of the ad-hoc logging of Self.self
(which I’ll do in the next commit).