Don’t put timestamp in the log output
Most logging libraries (and hence likely most implementations of
LogHandler) print timestamps. This is a good thing since it gives users
freedom to choose how the timestamps should be formatted. However, it
means that if we include a timestamp in the string we pass to LogHandler
then users are likely to see duplicated timestamps.