Remove InternalLogHandler’s conformance to LogHandler
I want to deliberately create a split between the public LogHandler
interface and the internal interface used for logging in the SDK. The
motivation is that I’m going to want to change the signature of
logMessage(…) to add file name and line number information.

Also, now that InternalLogHandler does not extend another protocol, we
can use Sourcery to generate its mock and remove our manually-created
mock.