Use Fastlane’s trainer to generate JUnit report
It seems that xcpretty’s JUnit report generation doesn’t work. For
example, when we give it the xcodebuild output of a test run which had
lots of crashes [1], it completely ignores those tests and doesn’t
include them in its results.

So I’m switching to using trainer, which we already use in ably-cocoa.
It directly reads the .xcresult bundle emitted by Xcode and gives much
better results (for example, in [1] it correctly identifies the crashed
tests as failures).

Closes #513.

[1] https://github.com/ably/ably-asset-tracking-swift/issues/513#issuecomment-1349970753