Move the LogParser example app outside of the library’s directory
I want it to be easy for people to create a script that uses the
LogParser library. The easiest thing I can think of is to be able to
just copy and paste the existing example app and then make modifications
to it. This isn’t very easy with the current structure, since the
example app’s code and build config is a bit intertwined with that of
the LogParser library. By putting the example app in a separate
directory, we can make a copy of it by just duplicating the entire
directory.