Add a client for sdk-test-proxy
The code and comments are more or less copied from the Android codebase
at commit 7f05c3c. I’ve changed the class name from FaultProxyClient to
SDKTestProxyClient, as it’s more in line with the name of the proxy
repo.

We also add a test to check that the local proxy is running and that the
client is working. This means we now need to run the proxy in CI, so we
do that.

Resolves #537.