Add functionality for running parameterized tests
For our upcoming NetworkConnectivityTests, we want to run a predefined
list of test cases for all of the faults currently known to the SDK test
proxy server.

On Android, we use JUnit’s @Parameterized functionality to do this.
There isn’t a corresponding thing in XCTest, so here we build one.

The general approach (where to hook into the Objective-C runtime and
XCTest callbacks) is inspired by the implementation of the Quick testing
framework [1].

[1] https://github.com/Quick/Quick