Update Swift version in Package.swift to 5.7
And address the deprecation warning this introduces.

This should have been done alongside c918aea.

I believe that this compiler version change introduces a new version (2)
of Package.resolved, but calling `swift package resolve` alone doesn’t
force the compiler to re-generate this file – seems like it will only do
so next time we have an actual change in our resolved dependencies. So
I’ll let it perform that migration when it needs to in the future.