Log the result received from Blocking.run’s result handler
If the async operation itself contains a call to Blocking.run, then,
when the outer timeout elapses, the inner call to Blocking.run will fail
with an `interrupted` error. The result of this is that the outer
operation’s result handler will get called with this interrupted error
too. At the moment, though, the logs just say that the outer operation’s
result handler was called, and that can be confusing since people might
think that this means the outer operation succeeded, even though they’re
about to see another message telling them it timed out.