Extend `Blocking.run`’s interruption behaviour to work across threads
Another thing we need (well, "need" might be a stretch, let’s say a
"thing that solves a problem that’s entirely possibly of my own making";
see next paragraph) for the subscriber network connectivity tests.

I would not be at all surprised to find that there is a simpler way of
writing the tests that doesn’t require this vile thing, but this is the
solution I have, and it seems to work, and I don’t really feel like
investing more time in these tests.