Introduce TrackableState enum to match usage in Android
Now that we've freed up this name in 00d6870, we can use it correctly.
It is used in the public APIs of the SDK to refer to the state of a
_trackable_ (whereas ConnectionState is an internal type used to refer
to _Ably_ connection state).

This is a breaking change to the public API of the SDK. I’m comfortable
making this change since we haven’t yet done a non-preview release and I
think we are going to have more to come anyway.