Re-enable the SwiftLint trailing_whitespace rule
I’m not sure why we disabled this one in the first place; it avoids
spurious changes and gets rid of big red boxes in `git diff` output. It
might be that it's because (I vaguely believe) Xcode’s default settings
cause it to sometimes add trailing whitespace, but this is configurable,
and anyway even if not then it's just a matter of telling the linter to
fix things.