Enable multiline_arguments and multiline_arguments_brackets SwiftLint rules
I had to disable a couple because this rule doesn’t seem to handle
multiple trailing closures well. This is discussed in [1] but there
doesn’t seem to be a resolution — I tried the suggested
only_enforce_after_first_closure_on_first_line config option but that
seems cause it to allow rule violations in function calls that don’t
even have a trailing closure.

[1] https://github.com/realm/SwiftLint/issues/3605