Do not emit a resolution to subscriber for ABSENT or LEAVE presence messages
This mimics the behaviour of Android’s
updateForPresenceMessagesAndThenEmitStateEventsIfRequired at commit
4d13ae7.

I’ve only added this piecemeal change to get the Null* subscriber
network connectivity tests to pass. We’ll be removing this code soon
enough as part of our Android parity work, so I haven’t bothered
changing the tests since this part of the code is currently untested
anyway.