Add reusable unit tests for DefaultWorker
I’ll be using this in the unit tests for the upcoming
UpdatePublisherPresence worker.

I’m using the prefix performTest_ for the test methods instead of just
test_, because SwiftLint insists on trying to remove the `private`
modifier when autocorrecting the private_unit_test rule. Adding a
swiftlint:disable comment causes it to complain that there is no
violation to ignore (which makes sense because from what I can tell it’s
only meant to consider methods contained inside XCTestCase subclasses,
which DefaultWorkerTestScenarios is not), but then it surely shouldn’t
be trying to autocorrect it either. Seems like a SwiftLint bug to me.