Make SubscriberQueueWorkerProperties mockable
Our upcoming subscriber worker classes will be calling methods on
SubscriberQueueWorkerProperties instances. These methods have
non-trivial implementations, the behaviour of which we do not wish to
rely on as a way of verifying that a worker has invoked the correct
method on the properties. Hence we wish to be able to pass a mock
properties object to the worker.